feat: delete auth token at log out
