+++ src/app/not-found.tsx
@@ -1,19 +1,10 @@
 "use client";
 
+import { useSignOut } from "@/hooks/useSignOut";
 import { Button } from "@nextui-org/button";
-import { useNeynarContext } from "@neynar/react";
-import Cookies from "js-cookie";
-import { useRouter } from "next/navigation";
 
 export default function ForbiddenPage() {
-  const router = useRouter();
-  const { logoutUser } = useNeynarContext();
-
-  const handleRetrySignIn = () => {
-    Cookies.remove("pyv2_auth_token");
-    logoutUser();
-    router.push("/");
-  };
+  const signOut = useSignOut();
 
   return (
     <div className="text-center">
@@ -21,7 +12,7 @@ export default function ForbiddenPage() {
       <p className="mb-6">
         You do not have permission to access this page. Please sign in to continue.
       </p>
-      <Button color="primary" onPress={handleRetrySignIn}>
+      <Button color="primary" onPress={signOut}>
         Retry Sign In
       </Button>
     </div>
